edit: add confirmation before deleting note and protect password

+++ app/src/main/java/me/martichou/be/grateful/fragments/ShowEditFragment.kt
@@ -4,9 +4,11 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
+import com.google.android.material.snackbar.Snackbar
 import me.martichou.be.grateful.databinding.FragmentShoweditBinding
 import me.martichou.be.grateful.viewmodels.EditViewModel
 import me.martichou.be.grateful.viewmodels.getNotesRepository
@@ -37,16 +39,22 @@ class ShowEditFragment : Fragment() {
      * Delete the note button handler
      */
     fun deletethisnote(view: View) {
-        viewModel.deleteNote()
-        findNavController().navigate(ShowEditFragmentDirections.actionEditFragmentToMainFragment())
+        AlertDialog.Builder(requireContext()).setMessage("Do you really want to delete it?").setPositiveButton("Yes") { _, _ ->
+            viewModel.deleteNote()
+            findNavController().navigate(ShowEditFragmentDirections.actionEditFragmentToMainFragment())
+        }.setNegativeButton("No") { dialog, _ -> dialog.dismiss() }.show()
     }
 
     /**
      * Save the note edited button handler
      */
     fun editthisnote(view: View) {
-        viewModel.updateNote(binding.editnoteTitle.text.toString(), binding.editnoteContent.text.toString())
-        findNavController().popBackStack()
+        if (!binding.editnoteTitle.text.isNullOrEmpty()) {
+            viewModel.updateNote(binding.editnoteTitle.text.toString(), binding.editnoteContent.text.toString())
+            findNavController().popBackStack()
+        } else {
+            Snackbar.make(binding.root, "The title can't be null", Snackbar.LENGTH_SHORT)
+        }
     }
 
 }
